Extract classes from UberSuper for more OOP
