+++ src/main/java/Exceptions/UberExceptions.java
@@ -1,0 +1,9 @@
+package Exceptions;
+
+public class UberExceptions extends RuntimeException {
+  //exceptions from chatBot
+    public UberExceptions(String message) {
+      super(message);
+    }
+}
+
