+++ src/main/java/Tasks/Deadline.java
@@ -1,0 +1,29 @@
+package Tasks;
+
+import java.time.LocalDateTime;
+
+public class Deadline extends Task {
+    private final LocalDateTime deadLine;
+
+    public Deadline(String description, LocalDateTime deadLine) {
+        super(description, TaskType.DEADLINE);
+        this.deadLine = deadLine;
+    }
+
+    @Override
+    public String toString() {
+        return String.format("[%s][%s] %s %s",
+                TaskType.DEADLINE.getSymbol(),
+                done() ? "X" : "",
+                desc(),
+                "(by: " + display(this.deadLine) + ")");
+    }
+    @Override
+    public String formattedString() {
+        return String.format("%s | %d | %s | %s",
+                type().getSymbol(),
+                done() ? 1 : 0,
+                desc(),
+                this.deadLine.format(STORAGE_DATETIME));
+    }
+}
