+++ src/main/java/Tasks/TaskType.java
@@ -1,0 +1,19 @@
+package Tasks;
+
+public enum TaskType {
+    TODO("T"),
+    DEADLINE("D"),
+    EVENT("E");
+
+    private final String symbol;
+
+    TaskType(String symbol) {
+        this.symbol = symbol;
+    }
+
+    public String getSymbol() {
+        return symbol;
+    }
+
+}
+
