+++ src/main/java/Utils/Parser.java
@@ -1,0 +1,63 @@
+package Utils;
+
+import Exceptions.UberExceptions;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+
+public class Parser {
+    // ===== Date/Time parsing helpers =====
+    public LocalDateTime parseWhen(String raw) throws UberExceptions {
+        String s = raw.trim();
+
+        // 1) ISO date-time: 2019-12-02T18:00 (or "2019-12-02 18:00")
+        try {
+            if (s.contains("T")) return LocalDateTime.parse(s);
+            if (s.matches("\\d{4}-\\d{2}-\\d{2}\\s+\\d{2}:\\d{2}")) {
+                return LocalDateTime.parse(s.replace(' ', 'T'));
+            }
+        } catch (DateTimeParseException ignore) {}
+
+        // 2) ISO date only: 2019-12-02  (treat as 00:00)
+        try { return LocalDate.parse(s).atStartOfDay(); } catch (DateTimeParseException ignore) {}
+
+        // 3) dd/MM/yyyy HHmm   e.g. 2/12/2019 1800
+        try {
+            DateTimeFormatter f = DateTimeFormatter.ofPattern("d/M/uuuu HHmm");
+            return LocalDateTime.parse(s, f);
+        } catch (DateTimeParseException ignore) {}
+
+        // 4) dd/MM/yyyy        (00:00)
+        try {
+            DateTimeFormatter f = DateTimeFormatter.ofPattern("d/M/uuuu");
+            return LocalDate.parse(s, f).atStartOfDay();
+        } catch (DateTimeParseException ignore) {}
+
+        // 5) d-M-uuuu HHmm
+        try {
+            DateTimeFormatter f = DateTimeFormatter.ofPattern("d-M-uuuu HHmm");
+            return LocalDateTime.parse(s, f);
+        } catch (DateTimeParseException ignore) {}
+
+        // 6) d-M-uuuu (00:00)
+        try {
+            DateTimeFormatter f = DateTimeFormatter.ofPattern("d-M-uuuu");
+            return LocalDate.parse(s, f).atStartOfDay();
+        } catch (DateTimeParseException ignore) {}
+
+        throw new UberExceptions("I couldn't understand the date/time: \"" + raw + "\".\n"
+                + "Try formats like: 2019-12-02, 2019-12-02 18:00, 2/12/2019 1800.");
+    }
+
+    // Convenient display: if time is 00:00, show date only
+    private static String display(LocalDateTime dt) {
+        if (dt.toLocalTime().equals(LocalTime.MIDNIGHT)) {
+            return dt.toLocalDate().format(DISPLAY_DATE);
+        }
+        return dt.format(DISPLAY_DATETIME);
+    }
+
+}
